requirements.txt
pyserial
customtkinter
toml
requests
